# Forge Task/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel-News portal app",
        "version": "1.0.0"
    },
    "paths": {
        "/admin/users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get list of non-admin users",
                "description": "Returns a paginated list of non-admin users",
                "operationId": "users.index",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/admin/users/{user_id}/edit": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get user details for editing",
                "description": "Returns user data with favourites and comments",
                "operationId": "users.edit",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/admin/users/{user_id}/update": {
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update user details",
                "description": "Updates user name, email and password",
                "operationId": "updateUser",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID of the user to be updated",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "New name of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "New email of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "New password of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User updated successfully"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/admin/users/{user_id}/delete": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete user",
                "description": "Deletes user by ID",
                "operationId": "deleteUser",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID of the user to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User deleted successfully"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/admin/comments/{comment_id}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete comment",
                "description": "Deletes comment by ID",
                "operationId": "deleteComment",
                "parameters": [
                    {
                        "name": "comment_id",
                        "in": "path",
                        "description": "ID of the comment to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Comment deleted successfully"
                    },
                    "404": {
                        "description": "Comment not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/admin/favourites/{favourite_id}": {
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Delete favourite",
                "description": "Deletes favourite by ID",
                "operationId": "deleteFavourite",
                "parameters": [
                    {
                        "name": "favourite_id",
                        "in": "path",
                        "description": "ID of the favourite to be deleted",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Favourite deleted successfully"
                    },
                    "404": {
                        "description": "Favourite not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
        "/admin/logs": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get user log activity",
                "description": "Returns a list of user log activity",
                "operationId": "userLog",
                "responses": {
                    "200": {
                        "description": "User log activity retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                }
            }
        },
...
        "/password": {
            "put": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Update the user's password",
                "description": "Updates the user's password after validating the current password and the new password confirmation",
                "operationId": "updatePassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "current_password",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "current_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newsecret"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "newsecret"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "302": {
                        "description": "Successful operation",
                        "headers": {
                            "Location": {
                                "description": "The URL of the previous page with a status message",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "current_password": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The current password is incorrect."
                                                    }
                                                },
                                                "password": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The password confirmation does not match."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "required": [
                    "name",
                    "country",
                    "language",
                    "email",
                    "password",
                    "favourite_category"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier of the user.",
                        "type": "UUID",
                        "format": "string",
                        "readOnly": true
                    },
                    "name": {
                        "description": "The name of the user.",
                        "type": "string"
                    },
                    "country": {
                        "description": "The country of the user.",
                        "type": "string"
                    },
                    "language": {
                        "description": "The language of the user.",
                        "type": "string"
                    },
                    "email": {
                        "description": "The email address of the user.",
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "description": "The password of the user.",
                        "type": "string",
                        "format": "password",
                        "writeOnly": true
                    },
                    "favourite_category": {
                        "description": "The favourite category of the user.",
                        "type": "string"
                    },
                    "email_verified_at": {
                        "description": "The date and time when the email address of the user was verified.",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true,
                        "nullable": true
                    },
                    "userFavourites": {
                        "description": "The favourites of the user.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserFavourites"
                        },
                        "readOnly": true
                    },
                    "comments": {
                        "description": "The comments of the user.",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserComments"
                        },
                        "readOnly": true
                    }
                },
                "type": "object"
            },
            "UserComments": {
                "required": [
                    "user_id",
                    "comment"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier of the user comment.",
                        "type": "UUID",
                        "format": "string",
                        "readOnly": true
                    },
                    "news_id": {
                        "description": "The identifier of the user who made the comment.",
                        "type": "UUID",
                        "format": "string"
                    },
                    "user_id": {
                        "description": "The identifier of the news that has the comments.",
                        "type": "UUID",
                        "format": "string"
                    },
                    "comment": {
                        "description": "The content of the comment.",
                        "type": "string"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
            "UserFavourites": {
                "required": [
                    "user_id",
                    "title",
                    "description",
                    "newsURL",
                    "imageURL"
                ],
                "properties": {
                    "id": {
                        "description": "The unique identifier of the user favourite.",
                        "type": "UUID",
                        "format": "string",
                        "readOnly": true
                    },
                    "user_id": {
                        "description": "The identifier of the user who added the favourite.",
                        "type": "UUID",
                        "format": "string"
                    },
                    "title": {
                        "description": "The title of the favourite.",
                        "type": "string"
                    },
                    "description": {
                        "description": "The description of the favourite.",
                        "type": "string"
                    },
                    "newsURL": {
                        "description": "The URL of the news article related to the favourite.",
                        "type": "string",
                        "format": "uri"
                    },
                    "imageURL": {
                        "description": "The URL of the image related to the favourite.",
                        "type": "string",
                        "format": "uri"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    }
                },
                "type": "object"
            },
...
        }
    }
}